[PUR-33] Create chat space subscribe route
Added a new route /api/chat/spaces/subscribe to the application.
This API endpoint allows subscribing to Google Chat space events. It takes a list of space IDs as input and follows these steps:
Creates a Pub/Sub topic if it does not exist.
Creates a subscription for the topic.
Registers the space for workspace subscriptions to receive event notifications.

Change-Id: I08c50b6ffacd3131afb8c2e66ae6d6a58e9977f9