[PUR-38] Enhance dynamic tag generation in genrule for OCI image push
Modified the `genrule` for dynamic tag generation to accept two environment variables, `TAG_1` and `TAG_2`. If not defined, they default to `latest`. This allows for two tags to be applied when pushing the image to Artifact Registry: one tag is a timestamp for rollback purposes, and the other is `latest` for easy identification of the most recent version.

Change-Id: Icee7f195b1045cadcb2419b170fb1bac924fd10c